Finish project

+++ app/Http/Controllers/Author/PostController.php
@@ -8,12 +8,14 @@ use App\Category;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Notifications\NewAuthorPost;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Intervention\Image\Facades\Image;
 use Illuminate\Support\Facades\File;
 use Brian2694\Toastr\Facades\Toastr;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Notification;
 
 class PostController extends Controller
 {
@@ -97,7 +99,9 @@ class PostController extends Controller
 
         $post->categories()->attach($request->categories);
         $post->tags()->attach($request->tags);
-
+        //Notification
+        // $users = User::where('role_id', '1')->get();
+        // Notification::send($users, new NewAuthorPost($post));
         Toastr::success('Post Successfully Saved :)', 'Success');
         return redirect()->route('author.post.index');
     }
@@ -110,6 +114,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            Toastr::error('You are not authorized to access this post', 'Error');
+            return redirect()->back();
+        }
         return view('author.post.show', compact('post'));
     }
 
@@ -121,6 +129,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            Toastr::error('You are not authorized to access this post', 'Error');
+            return redirect()->back();
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit', compact('post', 'categories', 'tags'));
@@ -135,6 +148,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            Toastr::error('You are not authorized to access this post', 'Error');
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required',
             'image' => 'image',
@@ -205,6 +223,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            Toastr::error('You are not authorized to access this post', 'Error');
+            return redirect()->back();
+        }
+
         if (Storage::disk('public')->exists('post/' . $post->image)) {
             Storage::disk('public')->delete('post/' . $post->image);
         }
